Add a few points about Covid-19
These points are taken from the 36C3 telegram channel many of us at hackeriet have joined. 

We might also want to add this...

"We still have to be honest with ourselves when it comes to Covid. Every event poses a risk/reward calculation, and each of us will have to figure out for ourselves if the outcome of this calculation is within acceptable limits. At Hackeriet we'll try to improve the parameters for this calculation as well as we can, but please know that it's completely fine if you feel this isn't good enough for you, and choose not to attend in person. We'll miss you, and hope you'll join us online from a safe place instead!"